move code to delayed.js file
